fix: 유저 삭제 시 userId 사용

+++ src/main/kotlin/kr/co/shoppingcart/cart/api/user/UserController.kt
@@ -88,7 +88,7 @@ class UserController(
     ): ResponseEntity<Unit> {
         userUseCase.deleteUser(
             DeleteUserCommand(
-                currentUser.identificationValue,
+                currentUser.identificationValue.toLong(),
                 LoginProvider.valueOf(currentUser.provider),
             ),
         )
